+++ Src/ImplicitNullability.Plugin.Tests/test/data/UnitTests/Configuration/ImplicitNullabilityConfigurationEvaluatorTest/MultipleAttributes.cs
@@ -4,4 +4,4 @@
 [assembly: AssemblyMetadata("ImplicitNullability.AppliesTo", "OutParametersAndResult")]
 [assembly: AssemblyMetadata("ImplicitNullability.AppliesTo", "InputParameters")]
 
-[assembly: AssemblyMetadata("ImplicitNullability.AppliesTo", "InputParameters, RefParameters, OutParametersAndResult, Fields")]
+[assembly: AssemblyMetadata("ImplicitNullability.AppliesTo", "InputParameters, Fields")]